refactor(cli): improve manifest maintenance
The CLI manifest commands needed clearer structure and documentation for long-term upkeep.

- cmd/ppkgmgr/main.go: add file-level and version documentation comments.

- internal/cli/*.go: document commands, extract manifest helpers, and introduce a dedicated pkg updater workflow.

- internal/data/*.go: document manifest structures and helpers.

- internal/registry/*.go: document registry types and add package docs.

- pkg/req/*.go: add package documentation and annotate helper functions.
