+++ internal/cli/doc.go
@@ -1,0 +1,2 @@
+// Package cli implements the command-line interface for ppkgmgr.
+package cli
