+++ internal/cli/pkg.go
@@ -1,20 +1,12 @@
 package cli
 
 import (
-	"errors"
 	"fmt"
-	"io"
-	"os"
-	"path/filepath"
-	"strings"
-	"time"
-
-	"ppkgmgr/internal/data"
-	"ppkgmgr/internal/registry"
 
 	"github.com/spf13/cobra"
 )
 
+// newPkgCmd provides manifest management helpers under the `pkg` namespace.
 func newPkgCmd(downloader DownloadFunc) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "pkg",
@@ -29,6 +21,7 @@ func newPkgCmd(downloader DownloadFunc) *cobra.Command {
 	return cmd
 }
 
+// newPkgUpCmd installs the `pkg up` subcommand responsible for refreshing manifests.
 func newPkgUpCmd(downloader DownloadFunc) *cobra.Command {
 	var force bool
 	cmd := &cobra.Command{
@@ -46,145 +39,13 @@ func newPkgUpCmd(downloader DownloadFunc) *cobra.Command {
 	return cmd
 }
 
+// handlePkgUp executes the `pkg up` workflow using the cobra command context.
 func handlePkgUp(cmd *cobra.Command, downloader DownloadFunc, force bool) error {
-	stdout := cmd.OutOrStdout()
-	stderr := cmd.ErrOrStderr()
-
-	if downloader == nil {
-		fmt.Fprintln(stderr, "pkg up requires a downloader")
-		return cliError{code: 5}
+	updater := pkgUpdater{
+		downloader: downloader,
+		stdout:     cmd.OutOrStdout(),
+		stderr:     cmd.ErrOrStderr(),
+		force:      force,
 	}
-
-	root, err := storageDir()
-	if err != nil {
-		fmt.Fprintf(stderr, "failed to determine storage directory: %v\n", err)
-		return cliError{code: 5}
-	}
-
-	registryPath := filepath.Join(root, "registry.json")
-	store, err := registry.Load(registryPath)
-	if err != nil {
-		fmt.Fprintf(stderr, "failed to load registry: %v\n", err)
-		return cliError{code: 5}
-	}
-
-	if len(store.Entries) == 0 {
-		fmt.Fprintln(stdout, "no manifests registered")
-		return nil
-	}
-
-	var hadFailure bool
-
-	for i := range store.Entries {
-		entry := &store.Entries[i]
-		if entry.Source == "" || entry.LocalPath == "" {
-			fmt.Fprintf(stderr, "warning: skipping manifest with incomplete metadata: %s\n", displayValue(entry.Source))
-			hadFailure = true
-			continue
-		}
-
-		var previousTargets []manifestTarget
-		if manifestTargets, err := extractManifestTargets(entry.LocalPath); err == nil {
-			previousTargets = manifestTargets
-		} else if !errors.Is(err, os.ErrNotExist) {
-			fmt.Fprintf(stderr, "warning: failed to parse stored manifest %s: %v\n", displayValue(entry.Source), err)
-		}
-
-		changed, err := refreshStoredManifest(entry)
-		if err != nil {
-			fmt.Fprintf(stderr, "warning: failed to refresh %s: %v\n", displayValue(entry.Source), err)
-			hadFailure = true
-			continue
-		}
-		if !changed && !force {
-			fmt.Fprintf(stdout, "manifest unchanged: %s\n", displayValue(entry.Source))
-			continue
-		}
-		if changed {
-			fmt.Fprintf(stdout, "refreshed manifest: %s\n", displayValue(entry.Source))
-			if len(previousTargets) != 0 {
-				cleanupOldTargets(previousTargets, stderr)
-			}
-		} else {
-			fmt.Fprintf(stdout, "forced refresh: %s\n", displayValue(entry.Source))
-		}
-
-		if _, err := os.Stat(entry.LocalPath); err != nil {
-			fmt.Fprintf(stderr, "warning: manifest unavailable for %s: %v\n", displayValue(entry.Source), err)
-			hadFailure = true
-			continue
-		}
-
-		fd, err := data.Parse(entry.LocalPath)
-		if err != nil {
-			fmt.Fprintf(stderr, "warning: failed to parse manifest %s: %v\n", displayValue(entry.Source), err)
-			hadFailure = true
-			continue
-		}
-
-		if err := downloadManifestFiles(fd, downloader, stdout, stderr, false); err != nil {
-			hadFailure = true
-		} else {
-			fmt.Fprintf(stdout, "updated files for: %s\n", displayValue(entry.Source))
-		}
-	}
-
-	if err := store.Save(registryPath); err != nil {
-		fmt.Fprintf(stderr, "failed to save registry: %v\n", err)
-		return cliError{code: 5}
-	}
-
-	if hadFailure {
-		return cliError{code: 4}
-	}
-	return nil
-}
-
-func refreshStoredManifest(entry *registry.Entry) (bool, error) {
-	raw, err := data.LoadRaw(entry.Source)
-	if err != nil {
-		return false, err
-	}
-
-	if err := os.MkdirAll(filepath.Dir(entry.LocalPath), 0o755); err != nil {
-		return false, err
-	}
-
-	if err := os.WriteFile(entry.LocalPath, raw, 0o600); err != nil {
-		return false, err
-	}
-
-	_, computed, err := verifyDigest(entry.LocalPath, "")
-	if err != nil {
-		return false, err
-	}
-	changed := entry.UpdatedAt.IsZero() || !strings.EqualFold(entry.Digest, computed)
-	entry.Digest = computed
-	if entry.ID == "" {
-		entry.ID = generateEntryID(entry.Source)
-	}
-	if changed {
-		entry.UpdatedAt = time.Now().UTC()
-	}
-	return changed, nil
-}
-
-func extractManifestTargets(path string) ([]manifestTarget, error) {
-	fd, err := data.Parse(path)
-	if err != nil {
-		return nil, err
-	}
-	return manifestOutputPaths(fd)
-}
-
-func cleanupOldTargets(targets []manifestTarget, stderr io.Writer) {
-	for _, target := range targets {
-		if err := os.Remove(target.path); err != nil && !errors.Is(err, os.ErrNotExist) {
-			fmt.Fprintf(stderr, "warning: failed to remove outdated file %s: %v\n", target.path, err)
-		}
-	}
-}
-
-type manifestTarget struct {
-	path string
+	return updater.run()
 }
