+++ internal/data/doc.go
@@ -1,0 +1,2 @@
+// Package data parses and loads ppkgmgr manifest files.
+package data
