+++ internal/registry/doc.go
@@ -1,0 +1,2 @@
+// Package registry persists metadata about stored manifests.
+package registry
