+++ pkg/req/doc.go
@@ -1,0 +1,2 @@
+// Package req provides HTTP helpers for retrieving remote files.
+package req
